+++ src/Pegasus/Common/CIMMethodRep.h
@@ -98,7 +98,7 @@ public:
     // ATTN:
     Boolean existsQualifier(const String& name) const
     {
-	return (findQualifier(name) != -1) ? true : false;
+	return (findQualifier(name) != Uint32(-1)) ? true : false;
     }
 
     CIMQualifier getQualifier(Uint32 pos)
